requirements.txt
fastapi
uvicorn
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.3.4
langchain-google-genai
langchain-openai==0.2.5
langchain-pinecone
pinecone-client
pydantic
requests
python-dotenv

#Login & Database
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt==4.0.1

pypdf
pandas
PyGithub

beautifulsoup4
validators

# google drive
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

pymysql
cryptography

chromadb
langchain-chroma
langchain-text-splitters